@@ -3,9 +3,6 @@
   "version": "1.0.0",
   "description": "Project Catwalk",
   "main": "config.js",
-  "jest": {
-    "verbose": true
-  },
   "scripts": {
     "test": "jest",
     "dev": "next dev",
@@ -34,21 +31,28 @@
     "prop-types": "^15.7.2",
     "react": "^17.0.1",
     "react-bootstrap": "^1.5.0",
-    "react-dom": "^17.0.1"
+    "react-dom": "^17.0.1",
+    "react-image-magnify": "^2.7.4"
   },
   "devDependencies": {
     "@babel/core": "^7.13.1",
+    "@testing-library/jest-dom": "^5.11.9",
+    "@testing-library/react": "^11.2.5",
     "@wojtekmaj/enzyme-adapter-react-17": "^0.4.1",
     "babel-jest": "^26.6.3",
     "enzyme": "^3.11.0",
-    "eslint": "^7.20.0",
+    "eslint": "^7.21.0",
     "eslint-config-airbnb": "^18.2.1",
     "eslint-plugin-import": "^2.22.1",
     "eslint-plugin-jsx-a11y": "^6.4.1",
     "eslint-plugin-react": "^7.22.0",
     "eslint-plugin-react-hooks": "^4.2.0",
+    "identity-obj-proxy": "^3.0.0",
     "jest": "^26.6.3",
     "react-addons-test-utils": "^15.6.2",
     "react-test-renderer": "^17.0.1"
+  },
+  "directories": {
+    "test": "tests"
   }
 }
