@@ -1,19 +1,22 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import AddQuestion from './components/AddQuestion';
+import AddMore from './components/AddMore';
 import Questions from './components/Questions';
 import QASearchBar from './components/SearchBar';
 
 const QAs = ({ productId }) => (
   <div>
-    <QASearchBar />
+    <QASearchBar productId={productId} />
     <Questions productId={productId} />
-    <AddQuestion />
+    <AddMore productId={productId} />
   </div>
 );
 
 QAs.propTypes = {
-  productId: PropTypes.number.isRequired,
+  productId: PropTypes.oneOfType([
+    PropTypes.string,
+    PropTypes.number,
+  ]).isRequired,
 };
 
 export default QAs;
