@@ -4,14 +4,16 @@ import PropTypes from 'prop-types';
 import config from '../../../config';
 import Q from './Q';
 
+// eslint-disable-next-line no-unused-vars
 function Questions({ productId }) {
   const [questions, setQuestions] = useState([]);
+  const [render, setRender] = useState(false);
   const id = productId;
+  const count = 'count=4';
 
   const getQuestions = () => {
     const options = {
-      url: `https://app-hrsei-api.herokuapp.com/api/fec2/hr-bld/qa/questions?product_id=${id}`,
-
+      url: `https://app-hrsei-api.herokuapp.com/api/fec2/hr-bld/qa/questions?product_id=${id}&${count}`,
       method: 'get',
       headers: {
         Authorization: config.TOKEN,
@@ -25,17 +27,31 @@ function Questions({ productId }) {
 
   useEffect(() => {
     getQuestions();
-  }, []);
+    setRender(false);
+  }, [render]);
 
   return (
     <>
-      {questions.map((item) => <Q question={item} key={item.question_id} answers={item.answers} />)}
+      {questions.map((item) => (
+        <Q
+          question={item}
+          key={item.question_id}
+          answers={item.answers}
+          setRender={setRender}
+          id={id}
+          count={count}
+          setQuestions={setQuestions}
+        />
+      ))}
     </>
   );
 }
 
 Questions.propTypes = {
-  productId: PropTypes.number.isRequired,
+  productId: PropTypes.oneOfType([
+    PropTypes.string,
+    PropTypes.number,
+  ]).isRequired,
 };
 
 export default Questions;
