@@ -1,33 +1,16 @@
 /* eslint-disable no-console */
 import React, { useState, useEffect } from 'react';
 import PropTypes from 'prop-types';
-import axios from 'axios';
+import Rating from '@material-ui/lab/Rating';
+
 import ProductBreakdown from './ProductBreakdown';
 import RatingBreakdown from './RatingBreakdown';
-import Stars from './Stars';
-import config from '../../../config';
 
-const ReviewsBreakdown = ({ productId }) => {
-  const [productMeta, setProductMeta] = useState(null);
+const ReviewsBreakdown = ({
+  productMeta, setProductRating, selectedRatings, setSelectedRatings, setRatingsLength,
+}) => {
   const [rating, setRating] = useState(null);
   const [recommended, setRecommended] = useState(null);
-
-  const getProductMeta = (product) => {
-    const api = 'https://app-hrsei-api.herokuapp.com/api/fec2/hr-bld/reviews/meta';
-    const options = {
-      url: `${api}?product_id=${product}`,
-      method: 'get',
-      headers: {
-        Authorization: config.TOKEN,
-      },
-    };
-
-    axios(options)
-      .then((res) => {
-        setProductMeta(res.data);
-      })
-      .catch((err) => { console.log(err); });
-  };
 
   const ratingCreator = (ratingsObj) => {
     let allRatings = 0;
@@ -38,6 +21,7 @@ const ReviewsBreakdown = ({ productId }) => {
       allRatings += Number(keys[i]) * Number(values[i]);
       ratingCount += Number(values[i]);
     }
+    setProductRating((allRatings / ratingCount).toFixed(1));
     return setRating((allRatings / ratingCount).toFixed(1));
   };
 
@@ -51,32 +35,51 @@ const ReviewsBreakdown = ({ productId }) => {
   };
 
   useEffect(() => {
-    getProductMeta(productId);
+    ratingCreator(productMeta.ratings);
+    reviewPercentage(productMeta.ratings, productMeta.recommended.true);
   }, []);
-
-  useEffect(() => {
-    if (productMeta) {
-      ratingCreator(productMeta.ratings);
-      reviewPercentage(productMeta.ratings, productMeta.recommended.true);
-    }
-  }, [productMeta]);
 
   return (
     <div>
-      <div><h4>{rating}</h4></div>
-      <Stars rating={rating} />
+      <div className="avg-rating">{rating}</div>
+      <Rating className="star-rating" value={Number(rating)} precision={0.25} readOnly />
       <div>
         {recommended}
         % of reviews recommend this product
       </div>
-      <RatingBreakdown />
+      <RatingBreakdown
+        ratings={productMeta.ratings}
+        selectedRatings={selectedRatings}
+        setSelectedRatings={setSelectedRatings}
+        setRatingsLength={setRatingsLength}
+      />
       <ProductBreakdown />
     </div>
   );
 };
 
 ReviewsBreakdown.propTypes = {
-  productId: PropTypes.string.isRequired,
+  productMeta: PropTypes.shape({
+    ratings: PropTypes.shape({
+      1: PropTypes.string,
+      2: PropTypes.string,
+      3: PropTypes.string,
+      4: PropTypes.string,
+      5: PropTypes.string,
+    }),
+    recommended: PropTypes.shape({
+      true: PropTypes.string,
+    }),
+  }),
+  setProductRating: PropTypes.func.isRequired,
+  selectedRatings: PropTypes.arrayOf(PropTypes.number),
+  setSelectedRatings: PropTypes.func.isRequired,
+  setRatingsLength: PropTypes.func.isRequired,
+};
+
+ReviewsBreakdown.defaultProps = {
+  productMeta: null,
+  selectedRatings: PropTypes.arrayOf(null),
 };
 
 export default ReviewsBreakdown;
