@@ -1,4 +1,4 @@
-import React, { useState } from 'react';
+import React, { useState, useRef } from 'react';
 import ProductDescription from '../components/Description-Scott/ProductDescription';
 import Comparison from '../components/Comparison-Dorien/Comparison';
 import QAs from '../components/QAs-Malcolm/QAs';
@@ -6,13 +6,21 @@ import Reviews from '../components/Reviews-Jim/Reviews';
 
 const App = () => {
   const [productId, setProductId] = useState(18078);
+  const [productName, setProductName] = useState('Camo Onesie');
+  const [productRating, setProductRating] = useState(null);
+  const reviewsRef = useRef();
 
   return (
-    <div>
-      <ProductDescription productId={productId} />
+    <div className="App">
+      <ProductDescription
+        productId={productId}
+        productRating={productRating}
+        reviewsRef={reviewsRef}
+        setProductNameGlobal={setProductName}
+      />
       <Comparison productId={productId} setProductId={setProductId} />
-      <QAs productId={productId} />
-      <Reviews productId="18201" />
+      <QAs productId={productId} productName={productName} />
+      <Reviews productId="18201" reviewsRef={reviewsRef} setProductRating={setProductRating} productName={productName} />
     </div>
   );
 };
